dom/manifest/test/test_ManifestProcessor_JSON.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1079453
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1079453</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script src="common.js"></script>
  <script>
/**
 * JSON parsing/processing tests
 * https://w3c.github.io/manifest/#processing
 **/
'use strict';
var invalidJson = ['', ` \t \n ${whiteSpace} `, '{', '{[[}'];
invalidJson.forEach((testString) => {
  var expected = `Expect to recover from invalid JSON: ${testString}`;
  data.jsonText = testString;
  var result = processor.process(data);
  SimpleTest.is(result.start_url, docURL.href, expected);
});

var validButUnhelpful = ["1", 1, "", "[{}]", "null"];
validButUnhelpful.forEach((testString) => {
  var expected = `Expect to recover from invalid JSON: ${testString}`;
  data.jsonText = testString;
  var result = processor.process(data);
  SimpleTest.is(result.start_url, docURL.href, expected);
});
  </script>
</head>
